+++ lib/public/views/SimulationPasses/PerLhcPeriodOverview/SimulationPassesPerLhcPeriodOverviewPage.js
@@ -19,7 +19,7 @@ import { filtersPanelPopover } from '../../../components/Filters/common/filtersP
 import { estimateDisplayableRowsCount } from '../../../utilities/estimateDisplayableRowsCount.js';
 import { simulationPassesActiveColumns } from '../ActiveColumns/simulationPassesActiveColumns.js';
 import spinner from '../../../components/common/spinner.js';
-import { breadcrumps } from '../../../components/common/navigation/breadcrumps.js';
+import { breadcrumps } from '../../../components/common/navigation/breadcrumbs.js';
 import { contextualWarning } from '../../../components/common/contextualMessage/contextualMessage.js';
 
 const TABLEROW_HEIGHT = 42;
